Separate code into ES2015-modules
